Revert "make validate a shared helper"
This reverts commit 41763583e0b896f827349e479e23b4a89d3e8d63.
